# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Riffagro OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://lara.loc/",
            "description": "L5 Swagger Open localhost api server"
        },
        {
            "url": "https://projects.dev/api/v1",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
...
        "/api/get_current_user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get current user",
                "description": "Returns User data",
                "operationId": "getCurrentUser",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Not Authorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}